Tratamento de excessão com Optional e criação do delete.

+++ src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -14,6 +14,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.util.UriComponentsBuilder;
 
+
 import javax.persistence.EntityNotFoundException;
 import javax.transaction.Transactional;
 import javax.validation.Valid;
@@ -21,6 +22,7 @@ import java.net.URI;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/topicos") /*evitar repetir a URL em todos os métodos*/
@@ -51,6 +53,7 @@ public class TopicosController {
 
     //@RequestMapping(value = "/topicos",method = RequestMethod.POST)
     @PostMapping
+    @Transactional
     public ResponseEntity<TopicoDto> cadastrar(@RequestBody @Valid TopicoForm topicoform, UriComponentsBuilder uriBuilder) {
         Topico topico = topicoform.converter(cursoRepository);
         topicoRepository.save(topico);
@@ -62,22 +65,33 @@ public class TopicosController {
    // @RequestMapping(value = "/topicos/{id}", method = RequestMethod.GET)
     @GetMapping("/{id}")
     public ResponseEntity<DetalhesDoTopicoDto> detalhar(@PathVariable Long id) {
-        Topico topico = topicoRepository.getById(id);
-        return  ResponseEntity.ok(new DetalhesDoTopicoDto(topico));
+        Optional<Topico> topico = topicoRepository.findById(id);
+        if (topico.isPresent()){
+            return  ResponseEntity.ok(new DetalhesDoTopicoDto(topico.get()));
+        }
+        return ResponseEntity.notFound().build();
     }
 
     @PutMapping("/{id}")
     @Transactional
     public ResponseEntity<TopicoDto> atualizar(@PathVariable Long id, @RequestBody @Valid AtualizacaoTopicoForm form) {
-        Topico topico = form.atualizar(id, topicoRepository);
-        return ResponseEntity.ok(new TopicoDto(topico));
+       Optional<Topico> optional = topicoRepository.findById(id);
+       if (optional.isPresent()) {
+           Topico topico = form.atualizar(id, topicoRepository);
+           return ResponseEntity.ok(new TopicoDto(topico));
+       }
+       return ResponseEntity.notFound().build();
     }
 
    @DeleteMapping("/{id}")
    @Transactional
-   public ResponseEntity<TopicoDto> deletar(@PathVariable Long id){
-    topicoRepository.deleteById(id);
-    return ResponseEntity.ok(null);
+   public ResponseEntity<?> remover(@PathVariable Long id){
+    Optional<Topico> optional = topicoRepository.findById(id);
+    if (optional.isPresent()){
+        topicoRepository.deleteById(id);
+        return ResponseEntity.ok().build();
+    }
+    return ResponseEntity.notFound().build();
    }
